+++ src/main/resources/templates/login/before/signup.html
@@ -15,7 +15,6 @@
         <a th:href="@{/main}"><img class="top-logo" th:src="@{img/Logo.png}" alt=" 로고"></a>
         <div class="loginbar">
             <a th:href="@{/login}">로그인</a> |
-            <a th:href="@{/signup}">회원가입</a>
         </div>
     </div>
 </header>
